Feature/team matching (#113)
This PR sets up a route and controller to match teams based on roles
required per team, and common overlapping availability.

The values for roles required and minimum overlapping hours and days per
week can be found and adjusted in the `team-matching-helpers.js` file as
so:
```
const sweRequired = 3;
const uxRequired = 2;
const pmRequired = 1;

const minumumDaysOverlapRequired = 3;
const minimumHoursOverlapRequired = 8;
```

The route to call is
`POST .../generate-team`
It optionally accepts a json body object with property
`startingMembersIds` which expects an array of User ids that can start
off a project. This is optional, and the controller will return an error
if the users included are already assigned to a project, or if the
provided users supersede the required role count for a given role.
Request body would look like:
```
{
     "startingMembersIds" : ["6543138038382", "1384938592092"]
}
```

Again, starting members are optional, and when not provided, the
controller will craft a new team from scratch.

The route also accepts query parameters `count` and `offest`. By
default, the controller will fetch 50 members of each role to attempt to
match a team. Providing a custom `count` value means the controller will
fetch `[count]` number of users of each role to attempt to match a team.
The offset parameter can be used if `x` count was attempted, and no team
was possible, the provided offset value will get the _next_ `[count]`
number of users, starting at entry `[offset]` from the database,
essentially getting the next set of users to attempt a team match.

A successful response will look something like:
```
{
    "team": [
        "Ulises Jakubowski: Software Engineer",
        "Mayra Stiedemann: Software Engineer",
        "Terrence Erdman: Software Engineer",
        "Gavin Cummerata: UX Designer",
        "Jeanie Hilll: UX Designer",
        "Allison McCullough: Product Manager"
    ],
    "project": "65ced8db8b81f76bc9592a95",
    "totalCommonHours": 37,
    "commonAvailability": {
        "SUN": [
            [
                "5:30 PM",
                "8:30 PM"
            ]
        ],
        "MON": [
            [
                "12:00 PM",
                "5:00 PM"
            ]
        ],
        "WED": [
            [
                "12:00 PM",
                "5:30 PM"
            ]
        ],
        "THU": [
            [
                "12:00 PM",
                "5:00 PM"
            ]
        ]
    }
}
```


This PR also sets up the backend to support the Product Manager role. 
There are lines of code in `utils/seed/seedDatabase.js` and
`utils/seed/utils/project.js` that are commented out because they break
the frontend. TODOs are included to uncomment when the frontend is set
up to handle product managers.